change the package structure to use the src layout
