+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-description-file = README.md
+description_file = README.md
 
 [tool:pytest]
 addopts = --ignore=scripts
@@ -15,7 +15,7 @@ max-line-length = 88
 docstring-convention = numpy
 per-file-ignores =
     __init__.py: F401
-    pyradiosky/tests/*:D
+    tests/*:D
     docs/*:D,A
     setup.py:D
 rst-roles =
