+++ src/pyradiosky/data/fhd_catalog_with_beam_values.sav
