+++ src/pyradiosky/data/gsm_galactic.skyh5
