+++ src/pyradiosky/data/gsm_icrs.skyh5
