+++ src/pyradiosky/data/healpix_disk.skyh5
