+++ src/pyradiosky/data/mock_hera_text_2458098.27471.txt
