+++ src/pyradiosky/data/old_skyh5_point_sources.skyh5
