+++ src/pyradiosky/data/pointsource_catalog.txt
